chore(listItem): send email notification when a list item is unpublished manually
