+++ src/server/components/dashboard/listsItems/item/update/actionHandlers/helpers.ts
@@ -1,9 +1,10 @@
 import { createListSearchBaseLink } from "server/components/lists";
 import { getListItemContactInformation } from "server/models/listItem/providers/helpers";
 import serviceName from "server/utils/service-name";
-import { sendDataPublishedEmail } from "server/services/govuk-notify";
+import { sendDataPublishedEmail, sendManualUnpublishedEmail } from "server/services/govuk-notify";
 import { ListItemWithAddressCountry } from "server/models/listItem/providers/types";
-
+import type { List } from "server/models/types";
+import { lowerCase, startCase } from "lodash";
 export async function sendPublishedEmail(listItem: ListItemWithAddressCountry) {
   const searchLink = createListSearchBaseLink(listItem.type);
   const { contactName, contactEmailAddress } = getListItemContactInformation(listItem);
@@ -16,4 +17,17 @@ export async function sendPublishedEmail(listItem: ListItemWithAddressCountry) {
     listItem.address.country.name,
     searchLink
   );
+}
+
+export async function sendUnpublishEmail(list: List) {
+  if (list?.jsonData?.users) {
+    const tasks = list.jsonData.users.map(async (user) => {
+      await sendManualUnpublishedEmail({
+        emailAddress: user,
+        serviceType: lowerCase(startCase(list.type)),
+        country: list.country?.name as string,
+      });
+    });
+    await Promise.allSettled(tasks);
+  }
 }
