+++ src/server/config/server-config.ts
@@ -33,7 +33,7 @@ export const NOTIFY = {
     editAnnualReviewDate: process.env.GOVUK_NOTIFY_EDIT_ANNUAL_REVIEW_DATE_TEMPLATE_ID?.trim() ?? "",
     newListItemSubmitted: process.env.GOVUK_NOTIFY_NEW_LIST_ITEM_SUBMISSION_TEMPLATE_ID?.trim() ?? "",
     editProviderDetails: process.env.GOVUK_NOTIFY_EDIT_PROVIDER_DETAILS_TEMPLATE_ID?.trim() ?? "",
-
+    listItemUnpublished: process.env.GOVUK_NOTIFY_ITEM_UNPUBLISHED_TEMPLATE_ID?.trim() ?? "",
     annualReviewNotices: {
       postOneMonth: process.env.GOVUK_NOTIFY_ANNUAL_REVIEW_POST_ONE_MONTH_NOTICE?.trim() ?? "",
       postOneWeek: process.env.GOVUK_NOTIFY_ANNUAL_REVIEW_POST_ONE_WEEK_NOTICE?.trim() ?? "",
