+++ src/server/services/govuk-notify.ts
@@ -184,7 +184,7 @@ export async function sendAnnualReviewCompletedEmail(
   }
 }
 
-export async function sendProviderChangedDetailsEmail({
+export async function sendManualUnpublishedEmail({
   emailAddress,
   serviceType,
   country,
@@ -200,34 +200,21 @@ export async function sendProviderChangedDetailsEmail({
     }
 
     const personalisation = {
-      typeSingular: pluralize.singular(serviceType),
+      type: pluralize.singular(serviceType),
       country,
     };
 
     logger.info(
-      `personalisation for sendProviderChangedDetailsEmail: ${JSON.stringify(personalisation)}, API key ${
+      `personalisation for sendManualUnpublishedEmail: ${JSON.stringify(personalisation)}, API key ${
         NOTIFY.apiKey
       }, email address ${emailAddress}`
     );
-    await getNotifyClient().sendEmail(NOTIFY.templates.editProviderDetails, emailAddress, {
+    await getNotifyClient().sendEmail(NOTIFY.templates.listItemUnpublished, emailAddress, {
       personalisation,
       reference: "",
     });
   } catch (error) {
-    logger.error(`The provider changed details email could not be sent due to error: ${(error as Error).message}`);
-  }
-}
-
-export async function sendProviderChangeDetailsEmailToAdmins(list: Pick<List, "jsonData" | "country" | "type">) {
-  if (list?.jsonData?.users) {
-    const tasks = list.jsonData.users.map(async (user) => {
-      await sendProviderChangedDetailsEmail({
-        emailAddress: user,
-        serviceType: lowerCase(startCase(list.type)),
-        country: list.country?.name as string,
-      });
-    });
-    await Promise.allSettled(tasks);
+    logger.error(`The provider manual unpublish email could not be sent due to error: ${(error as Error).message}`);
   }
 }
